refactor(api): move NonLlmServeApiName to data-plane (routing concept, not repo concept)
NonLlmServeApiName lived in repo/types.ts because of one convenient
Exclude<PerformanceApiName, NonLlmServeApiName> usage in the LLM executor,
but the type describes a routing decision (which API names bypass the LLM
source-routing graph) — it has no analogue to PerformanceApiName, which
genuinely is a repo concept (a column-domain enum on the performance
table).

Move it to data-plane/shared/api-names.ts and keep PerformanceApiName in
repo/types.ts where it belongs. Three consumers update their import:
sources/execute.ts and targets/emit.ts (Exclude<> usages) and
internal-debug-error.ts (DebugSourceApi union). A dedicated module rather
than co-locating with passthrough-serve avoids a circular type-only import
between passthrough-serve and internal-debug-error.
